# indiauthors-platform/packages/contracts/openapi/indiauthors-platform.yaml
openapi: 3.1.0
info:
  title: Indie Authors Platform API
  version: 0.0.1
  description: API boundary for indiauthors.com account, commerce, licensing, downloads, support, and admin domains.
servers:
  - url: https://api.indiauthors.com
paths:
...
  /v1/auth/login/start:
    post:
      summary: Start provider login flow by obtaining provider authorization URL
      operationId: startAuthLogin
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [returnUrl]
              properties:
                returnUrl:
                  type: string
                  format: uri
      responses:
        "201":
          description: Login flow started
          content:
            application/json:
              schema:
                type: object
                required: [authorizeUrl]
                properties:
                  authorizeUrl:
                    type: string
                    format: uri
        "503":
          description: Provider integration required
...
  /v1/account/summary:
    get:
      summary: Get authenticated account summary
      operationId: getAccountSummary
      responses:
        "200":
          description: Account summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AccountSummary"
        "401":
          description: Authentication required
        "503":
          description: Auth integration not configured
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: ia_session
  schemas:
...
    AccountSummary:
      type: object
      properties:
        error:
          type: string
        summary:
          type: string
security:
  - sessionCookie: []